requirements.txt
django-clearcache
django-extensions
djangorestframework
groq
channels
better-profanity
afinn
nltk
bleach
markdown
numpy
matplotlib
uvicorn[standard]